+++ packages/react-magma-dom/src/i18n/interface.ts
@@ -94,6 +94,17 @@ export interface I18nInterface {
       keyboardInstructionsTooltip: string;
       legendButtonAriaLabel: string;
     };
+    toolbar?: {
+      defaultTitle?: string;
+      downloadAsCsv?: string;
+      downloadAsJpg?: string;
+      downloadAsPng?: string;
+      exitFullScreen?: string;
+      makeFullScreen?: string;
+      moreOptionsAriaLabel?: string;
+      showAsTableTooltip?: string;
+      tabularRepresentationLabel?: string;
+    };
   };
   combobox: {
     clearIndicatorAriaLabel: string;
@@ -134,6 +145,7 @@ export interface I18nInterface {
     day: string;
     month: string;
     year: string;
+    invalidYearError: string;
 
     helpModal: {
       header: string;